requirements.txt
gradio
openai>=1.0.0
python-dotenv
pandas